pyproject.toml
[project]
name = "alphaforge"
version = "0.1.1"
description = "Local-first dashboard for managing RealTest trading strategies."
authors = [{ name = "AlphaForge Team" }]
dependencies = [
    "sqlalchemy>=2.0",
    "alembic",
    "pydantic>=2.0",
    "pydantic-settings",
    "pyyaml",
    "pyarrow",
    "pandas",
    "streamlit",
    "matplotlib",
    "typer>=0.9",
    "rich",
    "click",
    "plotly",
    "pystray>=0.19.3",
    "Pillow>=10.2.0",
    "pywin32>=306 ; sys_platform == 'win32'",
]
requires-python = ">=3.11"

[project.optional-dependencies]
build = [
    "pyinstaller>=6.5.0",
]

[project.scripts]
alphaforge = "alphaforge.ingestion.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alphaforge"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
